better separate main and ui
* main handles initialization from platform/file system
* ui handles launching iced
