+++ crates/ui/Cargo.toml
@@ -7,10 +7,8 @@ authors = [ "Dan Knutson <dan.knutson@gmail.com>" ]
 [dependencies]
 anyhow = { version = "1.0.66", features = ["backtrace"] }
 camino = "1.1.1"
-directories = "4.0.1"
 flume = { version = "0.10.14" }
 log = { version = "0.4", features = ["release_max_level_info"] }
-pretty_env_logger = "0.4"
 thiserror = "1.0.37"
 
 iced_style = "0.8.0"
