+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![deny(missing_debug_implementations)]
 #![forbid(unsafe_code)]
 
+pub mod config;
 pub mod logging;
 
 #[cfg(test)]
